addressing comments + moving application up/down checks
