+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/api/SubjectType.kt
@@ -2,5 +2,5 @@ package com.netflix.spinnaker.keel.api
 
 // currently we have only resource and costarain types
 enum class SubjectType {
-  RESOURCE, CONSTRAIN;
+  RESOURCE, CONSTRAINT;
 }
